+++ Sources/Sentry/include/SentryUIViewControllerPerformanceTracker.h
@@ -4,7 +4,9 @@
 
 @class SentrySpan;
 @class SentryInAppLogic;
+@class SentryTimeToDisplayTracker;
 @class UIViewController;
+@class SentryTracer;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -30,7 +32,7 @@ static NSString *const SENTRY_UI_PERFORMANCE_TRACKER_TTD_TRACKER
 
 @property (nonatomic, strong) SentryInAppLogic *inAppLogic;
 
-@property (nonatomic) BOOL enableWaitForFullDisplay;
+@property (nonatomic) BOOL alwaysWaitForFullDisplay;
 
 /**
  * Measures @c controller's @c loadView method.
@@ -100,6 +102,10 @@ static NSString *const SENTRY_UI_PERFORMANCE_TRACKER_TTD_TRACKER
                            callbackToOrigin:(void (^)(void))callback;
 
 - (void)reportFullyDisplayed;
+
+- (nullable SentryTimeToDisplayTracker *)startTimeToDisplayTrackerForScreen:(NSString *)screenName
+                                                         waitForFullDisplay:(BOOL)waitForFullDisplay
+                                                                     tracer:(SentryTracer *)tracer;
 
 @end